LayoutTests/imported/w3c/web-platform-tests/css/css-font-loading/fontface-fonts-loading-expected.txt

Font loading test

PASS document.fonts.ready is replaced as new fonts are loaded
